feat: calculate html view height
