--- ios-tripbook/View/TravelNews/TravelNewsDetail/WebView.swift
@@ -1,65 +1,0 @@
-//
-//  WebView.swift
-//  ios-tripbook
-//
-//  Created by RED on 2023/12/07.
-//
-
-import SwiftUI
-import WebKit
-
-//struct HTMLView: UIViewRepresentable {
-//    let text: String
-//
-//    func makeUIView(context: Context) -> WKWebView {
-//        return WKWebView()
-//    }
-//
-//    func updateUIView(_ uiView: WKWebView, context: Context) {
-//        uiView.loadHTMLString(text, baseURL: nil)
-//    }
-//}
-
-//import SwiftUI
-//import WebKit
-
-struct HTMLView: UIViewRepresentable {
-    let htmlString: String
-    @Binding var contentHeight: CGFloat
-    
-    func makeUIView(context: Context) -> WKWebView {
-        let webView = WKWebView()
-        webView.navigationDelegate = context.coordinator
-        return webView
-    }
-    
-    func updateUIView(_ uiView: WKWebView, context: Context) {
-        uiView.loadHTMLString(htmlString, baseURL: nil)
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        return Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, WKNavigationDelegate {
-        var parent: HTMLView
-        
-        init(_ parent: HTMLView) {
-            self.parent = parent
-        }
-        
-        func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-            webView.evaluateJavaScript("document.readyState", completionHandler: { (complete, error) in
-                if complete != nil {
-                    webView.evaluateJavaScript("document.body.scrollHeight", completionHandler: { (height, error) in
-                        if let height = height as? CGFloat {
-                            DispatchQueue.main.async {
-                                self.parent.contentHeight = height
-                            }
-                        }
-                    })
-                }
-            })
-        }
-    }
-}
